Bring LLVM backend to using LLVM 3.3 (pre-release)
This patch updates the LLVM back-end to use 3.3 (but fixes interface changes
with #ifdefs in case we want it to work with earlier versions. The #ifdefs can
be removed in the future if they turn out to be irrelevant). Note that a header
file we were using to detect an installed LLVM in third-party/llvm/install has
been moved, so now we use a different file for such detection.

It also allows the LLVM backend to produce Chapel programs with GASNet.  To do
so, we have slightly modified codegen_makefile and added an option to skip
program compilation in Makefile.exe, because the LLVM backend has already
produced the program - it just needs to build the launcher.

Lastly, the minor changes to runtime/include/qio/qio.h allow compilation with
clang without warnings, and util/chplenv/third-party-pkgs now allows colons in
the environment variables it works with (a bug fix to prevent future problems
rather than because anything relies upon it).

Reviewed by Tom.

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21366 3a8e244f-b0f2-452b-bcba-4c88e055c3ca