LLVM backend bug fixes, wide pointer optimizations, initial TBAA
[Reviewed by Tom, then bug fixes were added]

This patch includes two main improvements to the LLVM back-end. First, it
includes a new optimization flag --llvm-wide-opt (fLLVMWideOpt). When that
flag is given to the compiler, instead of working with wide pointers using the
usual runtime calls, the code generator will output LLVM pointers in a
different address space (currently 100) for wide pointers. Then, put and get
calls are represented in the LLVM IR as load, store, or memcpy. That way, the
normal LLVM optimizations can optimize wide pointer operations. For example,
the Loop Invariant Code Motion optimization might pull a repeated get
(represented as a load of an address space 100 pointer) out of a loop if it
can prove it is safe to do so. After LLVM optimizations have run, our own
transformation pass llvmGlobalToWide.cpp lowers the operations on address
space 100 pointers back to Chapel runtime calls and translates all types back
to the wide versions of those types (ie removes address space 100 and switches
to packed pointers).  Note that the llvmGlobalToWide optimization adds calls
to the Chapel runtime based upon function name and assumes that the symbols
are available in the linked runtime. It works that way because otherwise LLVM
optimizations will remove these functions as being unused before they need to
be added (late in optimization) in the global-to-wide pass.

The second improvement to the LLVM back-end is to add TBAA (type-based alias
analysis) metadata in order to improve the efficacy of the LLVM optimizations.
However, since the TBAA metadata generation needs more testing, it is disabled
with a #define.

Other bug fixes:
 - moved chpl_config_has_value/chpl_config_get_value to be functions and not
   inline, because they were causing warnings with GASNet builds and an
   earlier version of this patch; since they are not performance sensitive it
   probably doesn't make sense to inline them anyway).
 - cleaned up loading up INT64_MIN/MAX/etc and now also support INT_MAX when
   populating the LayeredValueTable with C constants. This enables the
   LLVM backend to be used with qthreads.
 - (for debugging purposes) support packed pointers in the C backend.
    All that was missing was to generate types and zero-initization correctly.
    Supporting packed pointers in the C backend helped to debug problems
    with them.
 - cleaned code generation for PRIM_WIDE_GET_{LOCALE,NODE,SUBLOC} and
   make them work correctly with packed pointers.
 - fix some cases in externCResolve.cpp where '\0' was being passed to a
   string argument (by replacing '\0' with NULL). It was a compiler warning
   when building with clang.
 - LLVM code generation should always add alloca instructions to the
   function entry point. Prior to this patch, we were adding alloca
   instructions possibly in loops (a bad idea - leads to stack overflow).

The LLVM backend remains experimental.

Thanks to UMD student Joe Yan for his work towards an earlier version of this
patch.

Also, in case you are wondering... compiler/util/llvm-global-to-wide is a
directory storing appropriate files to build the LLVM Global to Wide
transformation independently from the Chapel compiler and includes some unit
tests for it within the LLVM testing framework.

To use the functionality introduced by this patch, try something like
export CHPL_COMM=gasnet
export CHPL_WIDE_POINTERS=node16
make
chpl --llvm --llvm-wide-opt --fast some-program.chpl

Performance is only improved in limited cases with this patch, since the
wide pointer optimization adds calls that really should be inlined... and
getting those to inline is future work. However, the communication counts
should be lower if the LLVM wide pointer optimizations were helpful.

git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21431 3a8e244f-b0f2-452b-bcba-4c88e055c3ca